+++ src/controllers/ForumController.js
@@ -5,7 +5,9 @@ const RequestValidator = require("../utils/request");
 
 const getAllForums = async (req, res) => {
   try {
-    const forums = await ForumService.getAllForums();
+    const forums = await ForumService.getAllForums({
+      order: [["createdAt", "DESC"]],
+    });
     res.status(forums.code).json(forums);
   } catch (error) {
     console.log(error);
@@ -27,6 +29,36 @@ const getForumById = async (req, res) => {
   }
 };
 
+const getCurrentUserForums = async (req, res) => {
+  try {
+    const forums = await ForumService.getAllForums({
+      where: { userId: req.userId },
+      order: [["createdAt", "DESC"]],
+    });
+    res.status(forums.code).json(forums);
+  } catch (error) {
+    console.log(error);
+    res
+      .status(500)
+      .json(new ResponseClass.ErrorResponse(500, "Internal Server Error"));
+  }
+};
+
+const getForumsByUserId = async (req, res) => {
+  try {
+    const forums = await ForumService.getAllForums({
+      where: { userId: req.params.id },
+      order: [["createdAt", "DESC"]],
+    });
+    res.status(forums.code).json(forums);
+  } catch (error) {
+    console.log(error);
+    res
+      .status(500)
+      .json(new ResponseClass.ErrorResponse(500, "Internal Server Error"));
+  }
+};
+
 const createForum = async (req, res) => {
   try {
     if (!req.files) {
@@ -38,7 +70,7 @@ const createForum = async (req, res) => {
     body.userId = req.userId;
     const validate = RequestValidator.verifyRequest(body, [
       "title",
-      "content",
+      "description",
       "userId",
     ]);
 
@@ -77,10 +109,12 @@ const updateForum = async (req, res) => {
             )
           );
       }
-      const filename = ImageService.getFilename(forum.data.image);
-      const deleteImage = await ImageService.deleteFromGcs(filename);
-      if (deleteImage.code !== 200 && deleteImage.code !== 404) {
-        return res.status(deleteImage.code).json(deleteImage);
+      if (forum.data.image !== null) {
+        const filename = ImageService.getFilename(forum.data.image);
+        const deleteImage = await ImageService.deleteFromGcs(filename);
+        if (deleteImage.code !== 200 && deleteImage.code !== 404) {
+          return res.status(deleteImage.code).json(deleteImage);
+        }
       }
       const uploadImage = await ImageService.uploadToGcs(req.files[0], "Forum");
       if (uploadImage.code !== 200) {
@@ -113,10 +147,12 @@ const deleteForum = async (req, res) => {
             )
           );
       }
-      const filename = ImageService.getFilename(forum.data.image);
-      const deleteImage = await ImageService.deleteFromGcs(filename);
-      if (deleteImage.code !== 200 && deleteImage.code !== 404) {
-        return res.status(deleteImage.code).json(deleteImage);
+      if (forum.data.image !== null) {
+        const filename = ImageService.getFilename(forum.data.image);
+        const deleteImage = await ImageService.deleteFromGcs(filename);
+        if (deleteImage.code !== 200 && deleteImage.code !== 404) {
+          return res.status(deleteImage.code).json(deleteImage);
+        }
       }
     }
     const deleteForum = await ForumService.deleteForum(
@@ -135,20 +171,15 @@ const deleteForum = async (req, res) => {
 const createForumComment = async (req, res) => {
   try {
     const body = req.body;
-    body.userId = req.userId;
-    const validate = RequestValidator.verifyRequest(body, [
-      "comment",
-      "userId",
-    ]);
+    body.UserId = req.userId;
+    body.ForumId = req.params.id;
+    const validate = RequestValidator.verifyRequest(body, ["comment"]);
 
     if (validate !== true) {
       return res.status(400).json(validate);
     }
 
-    const createForumComment = await ForumService.createForumComment(
-      req.params.id,
-      body
-    );
+    const createForumComment = await ForumService.createForumComment(body);
     res.status(createForumComment.code).json(createForumComment);
   } catch (error) {
     console.log(error);
@@ -160,6 +191,8 @@ const createForumComment = async (req, res) => {
 
 const updateForumComment = async (req, res) => {
   try {
+    req.body.UserId = req.userId;
+    req.body.role = req.role;
     const updateForumComment = await ForumService.updateForumComment(
       req.params.id,
       req.body
@@ -175,6 +208,8 @@ const updateForumComment = async (req, res) => {
 
 const deleteForumComment = async (req, res) => {
   try {
+    req.body.UserId = req.userId;
+    req.body.role = req.role;
     const deleteForumComment = await ForumService.deleteForumComment(
       req.params.id
     );
@@ -190,6 +225,8 @@ const deleteForumComment = async (req, res) => {
 module.exports = {
   getAllForums,
   getForumById,
+  getCurrentUserForums,
+  getForumsByUserId,
   createForum,
   updateForum,
   deleteForum,
