+++ src/controllers/MenuController.js
@@ -32,10 +32,22 @@ const deleteMenu = async (req, res) => {
     const permanent = req.query.permanent ? req.query.permanent : false;
     if (permanent) {
       const menu = await MenuService.getMenuById(req.params.id);
-      const filename = ImageService.getFilename(menu.data.image);
-      const deleteImage = await ImageService.deleteFromGcs(filename);
-      if (deleteImage.code !== 200 && deleteImage.code !== 404) {
-        return res.status(deleteImage.code).json(deleteImage);
+      if (menu.data == null) {
+        return res
+          .status(404)
+          .json(
+            new ResponseClass.ErrorResponse(
+              (code = 404),
+              (message = "Menu not found")
+            )
+          );
+      }
+      if (menu.data.image !== null) {
+        const filename = ImageService.getFilename(menu.data.image);
+        const deleteImage = await ImageService.deleteFromGcs(filename);
+        if (deleteImage.code !== 200 && deleteImage.code !== 404) {
+          return res.status(deleteImage.code).json(deleteImage);
+        }
       }
     }
     const deleteMenu = await MenuService.deleteMenu(req.params.id, permanent);
@@ -53,10 +65,22 @@ const updateMenu = async (req, res) => {
     if (req.files) {
       console.log("coek");
       const menu = await MenuService.getMenuById(req.params.id);
-      const filename = ImageService.getFilename(menu.data.image);
-      const deleteImage = await ImageService.deleteFromGcs(filename);
-      if (deleteImage.code !== 200 && deleteImage.code !== 404) {
-        return res.status(deleteImage.code).json(deleteImage);
+      if (menu.data == null) {
+        return res
+          .status(404)
+          .json(
+            new ResponseClass.ErrorResponse(
+              (code = 404),
+              (message = "Menu not found")
+            )
+          );
+      }
+      if (menu.data.image !== null) {
+        const filename = ImageService.getFilename(menu.data.image);
+        const deleteImage = await ImageService.deleteFromGcs(filename);
+        if (deleteImage.code !== 200 && deleteImage.code !== 404) {
+          return res.status(deleteImage.code).json(deleteImage);
+        }
       }
       const uploadImage = await ImageService.uploadToGcs(req.files[0], "Menu");
       if (uploadImage.code !== 200) {
